move to postgres 9.6

+++ src/main/java/com/evgeniy/transformers/controller/BattleFieldController.java
@@ -15,7 +15,7 @@ public class BattleFieldController {
     private final BattleFieldRepo repo;
 
     @DeleteMapping("/battleFields")
-    public ResponseEntity<Void> deleteByIds(@RequestParam List<Long> ids) {
+    public ResponseEntity<Void> deleteByIds(@RequestParam List<Integer> ids) {
         repo.deleteByIdIn(ids);
         return ResponseEntity.ok().build();
     }
