+++ src/main/java/com/evgeniy/transformers/controller/ModificationController.java
@@ -15,7 +15,7 @@ public class ModificationController {
     private final ModificationRepo repo;
 
     @DeleteMapping("/modifications")
-    public ResponseEntity<Void> deleteByIds(@RequestParam List<Long> ids) {
+    public ResponseEntity<Void> deleteByIds(@RequestParam List<Integer> ids) {
         repo.deleteByIdIn(ids);
         return ResponseEntity.ok().build();
     }
