+++ src/main/java/com/evgeniy/transformers/controller/WeaponController.java
@@ -15,7 +15,7 @@ public class WeaponController {
     private final WeaponRepo repo;
 
     @DeleteMapping("/weapons")
-    public ResponseEntity<Void> deleteByIds(@RequestParam List<Long> ids) {
+    public ResponseEntity<Void> deleteByIds(@RequestParam List<Integer> ids) {
         repo.deleteByIdIn(ids);
         return ResponseEntity.ok().build();
     }
