+++ src/main/java/com/evgeniy/transformers/restrepo/BattleFieldRepo.java
@@ -9,8 +9,8 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 @RepositoryRestResource(collectionResourceRel = "battleFields", path = "battleFields")
-public interface BattleFieldRepo extends PagingAndSortingRepository<BattleField, Long> {
+public interface BattleFieldRepo extends PagingAndSortingRepository<BattleField, Integer> {
     @Transactional
     @RestResource(exported = false)
-    void deleteByIdIn(List<Long> ids);
+    void deleteByIdIn(List<Integer> ids);
 }
